Create MIS.py
Task 4 MIS

An independent set is a set of vertices in a graph such that no two of which are connected by an edge. The problem of finding Maximum Independent Sets (MIS) is NP-hard. In this task, you will solve small instances of the MIS with gate and analog-based quantum computing: 

Create 4 graphs with 3, 5, 6, and 7 nodes. You can define the edges as you prefer. 
For each graph, find the MIS using the following methods: 
The Gate-Based QAOA  
The Quantum adiabatic algorithm using analog based quantum computing. You can check the documentation in Pulser and Blockade for references. 
Compared the results